HMW-73 Moved the summarizeAssessments function to the utils file.
