+++ app/client/src/utils/utils.js
@@ -298,6 +298,42 @@ function normalizeString(str: string) {
   return str.trim().toUpperCase();
 }
 
+// Summarizes assessment counts by the status of the provided fieldname.
+function summarizeAssessments(waterbodies: Array<Object>, fieldName: string) {
+  const summary = {
+    total: 0,
+    unassessed: 0,
+    'Not Supporting': 0,
+    'Fully Supporting': 0,
+    'Insufficient Information': 0,
+    'Not Assessed': 0,
+    'Not Applicable': 0,
+  };
+
+  // ids will contain unique assessment unit id's of each waterbody,
+  // to ensure we don't count a unique waterbody more than once
+  const ids = [];
+
+  waterbodies?.forEach((graphic) => {
+    const field = graphic.attributes[fieldName];
+    const { assessmentunitidentifier: id } = graphic.attributes;
+
+    if (!field || field === 'X') {
+      summary['Not Applicable']++;
+    } else {
+      if (ids.indexOf(id) === -1) {
+        ids.push(id);
+        if (field === 'Not Supporting' || field === 'Fully Supporting') {
+          summary.total++;
+        }
+        summary[field]++;
+      }
+    }
+  });
+
+  return summary;
+}
+
 export {
   chunkArray,
   containsScriptTag,
@@ -320,4 +356,5 @@ export {
   convertDomainCode,
   getSelectedCommunityTab,
   normalizeString,
+  summarizeAssessments,
 };
